+++ rust/apps/wallets/src/zcash.rs
@@ -1,5 +1,4 @@
 use alloc::string::String;
-
 use alloc::vec::Vec;
 
 use app_utils::impl_public_struct;
@@ -37,7 +36,6 @@ pub fn generate_sync_ur(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use alloc::string::ToString;
     use alloc::vec;
 
     #[test]